graph in form view
