+++ custom_reports/views/employee_performance_reports_views.xml
@@ -57,7 +57,7 @@
     </record>
 
     
-    <!-- tree -->
+    <!-- employee performance report tree -->
     <record model="ir.ui.view" id="employee_performance_report_view_tree">
         <field name="name">Employee Performance Report</field>
         <field name="model">custom_reports.employee_performance_report</field>
@@ -69,40 +69,14 @@
         </field>
     </record>
 
-     <!-- form -->
+    <!-- employee performance report form -->
     <record model="ir.ui.view" id="employee_performance_report_view_form">
         <field name="name">Employee Performance Report</field>
         <field name="model">custom_reports.employee_performance_report</field>
         <field name="arch" type="xml">
             <form>
                 <sheet>
-                    <widget name="web_ribbon" title="Archived" bg_color="bg-danger"/>             
-
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-    
-<!-- 
-    <record id="employee_performance_report_view_form_graph" model="ir.ui.view">
-        <field name="name">Employee Performance Report Graph Form</field>
-        <field name="model">custom_reports.employee_performance_report</field>
-        <field name="inherit_id" ref="custom_reports.employee_performance_report_view_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//notebook" position="inside">
-                <page string="Employee Performance Report Graph">
-                    <h2>Hey There</h2>
-                    <widget name="bh_sidebar"/>
-                
-                    
-                </page>
-            </xpath>
-        </field>
-    </record> -->
-
-    
-<!-- <group>
+                    <group>
                         <div class="oe_title">
                             <h1>Employee Performance Report</h1>
                         </div>
@@ -123,7 +97,7 @@
                     </group>
 
                     <notebook>
-                        <page string="Employee Performances Pivot">
+                        <page string="Pivot">
                             <field name="employee_performance_ids">
                                 <pivot>
                                     <field name="employee_id" type="row"/>
@@ -134,31 +108,23 @@
                             </field>
                         </page>
                     </notebook>
-
                 </sheet>
             </form>
         </field>
     </record>
-
     
-
+    <!-- employee performance graph form view -->
     <record id="employee_performance_report_view_form_graph" model="ir.ui.view">
         <field name="name">Employee Performance Report Graph Form</field>
         <field name="model">custom_reports.employee_performance_report</field>
         <field name="inherit_id" ref="custom_reports.employee_performance_report_view_form"/>
         <field name="arch" type="xml">
             <xpath expr="//notebook" position="inside">
-                <page string="Employee Performance Report Graph">
-                    <widget name = "bh_sidebar"/>
-                  
-                    
+                <page string="Graph">
+                    <widget name="employee_performance_graph"/>
                 </page>
             </xpath>
         </field>
-    </record>
-
-
- 
-  
+    </record> 
 
 </odoo>
